Update selinux dev branch will latest dev main branch. (#11)
* This is the latest fixed policies. (#10)

* Catch up the SELINUX specific branch to current DEV. (#8)

* Added GLIBC 2.18 'module'

link to the GNU source GIT repository. So that once you pull this repo, it will also grab the glibc repo for you with the correct version.

* Create necessary patch file for glibc

GLIBC patch file and instructions on how to patch it after you download the repo.

* Adjusted the way the installer works

And accounted for setting the permissions.

* Correct documentation formatting.

Too bad, github MD formatting is not that great.

* Added more obvious steps.

Do not assume steps are done, go down the process.

* Fixed conditional checks and documentation.

I don't like this due to known pitfalls, but the existing code is in this format. So consistency.

* Documentation changes.

Updated information and formatting.
Mention people in the appropriate perspective in the Thank you section.

* Update README.md

Formatting

* Corrected conditional checks

* Do not use the macros

Wrong file context.

* Ooops, wrong useradd function

Not supposed to create the extra SELINUX user.

* fcontext scoping changes

So, specificity changes the order of the way the file contexts are processed.

* Enforce permissions

Need to enforce permissions.

* Update README.md

* Clean up security policies

There were redundant entries and a macro that covered the correct and needed permission.

* Updated RPM package

Ready for release, an RPM package of the updated and cleaned up rules.

* Update README.md

Firewall section.

* Update README.md

Ooops, should keep it in original directory even if project name is different.

* Fix formatting issues

Fix formatting issues